backend/requirements.txt
fastapi
uvicorn[standard]
numpy
rasterio
rio-tiler
herbie-data
